more non template functions
