+++ sphinxsim/sph_simulation/simulation_builder/continuum_simulation_builder.cpp
@@ -1,7 +1,7 @@
 #include "continuum_simulation_builder.hpp"
 
 #include "base_simulation_builder.hpp"
-#include "constraint_builder.hpp"
+#include "constraint_builder.h"
 
 namespace SPH
 {
